Project3 - Amazon Product Recommendation System/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
streamlit>=1.28.0
plotly>=5.17.0
